requirements_ai.txt
fastapi
uvicorn
python-multipart
openai
openai-whisper
python-dotenv
pandas
gspread
google-auth
supabase
regex
